requirements.txt
pandas
scikit-learn
matplotlib
notebook
pyyaml
torch==1.4.0+cpu
torchvision==0.5.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html